relay: manage emergency reboot manually
ESPHome has a timeout on wifi and mqtt connection after which it will
hard reboot the device hoping it could recover from an issue if any.

In our case this interferes with the failsafe script so we disable those
`reboot_timeout`s and instead to it manually at the end of each failsafe
loops.